requirements.txt
pytest>=7.4.0
pytest-html>=4.0.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
requests>=2.31.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
